Initial framework prototype
The goal here was to refactor the original Dot to have better support
for the fish, which required some extensive changes. In the process of
doing this however, I noticed a number of potential optimizations we
could make to avoid having to re-initialize everything on shell startup
(e.g. using executables instead of inline-defined functions).

The organization of the framework itself will hopefully make it clearer
about where shell-specific logic lives. Fortunately the core framework
code will run in any POSIX-compliant shell.