Remove duplicate entries from plugins list
We thought we were already doing this, but we forgot that plugins which
are depended on by other plugins but don't have dependencies themselves
would appear in both the "dependencies" and "independent plugins" lists.

Thus we need to pass both lists into our magic `awk` de-duplicator to
remove all duplicates.