Include submodules in `git clone` in README
I'm setting up a new system and wanted to get started with a solid sds foundation. After copy-pasting the get-started commands I got this error when I tried to run `bin/install`: 

```sh
bin/install: No such file or directory
```

The install script references a file in `../frameworks` which is a git submodule. We need the submodule to be checked out as well when we clone.